+++ src/main/java/org/shorts/model/moves/Psyshock.java
@@ -11,7 +11,7 @@ public class Psyshock extends Move {
     }
 
     @Override
-    protected int getDefendingStat(Pokemon user, Pokemon target, Battle battle) {
+    protected double getDefendingStat(Pokemon user, Pokemon target, Battle battle) {
         return target.calculateDefense();
     }
 }
