+++ src/main/java/org/shorts/model/moves/SacredSword.java
@@ -11,7 +11,7 @@ public class SacredSword extends Move implements SlicingMove {
     }
 
     @Override
-    protected int getDefendingStat(Pokemon attacker, Pokemon defender, Battle battle) {
+    protected double getDefendingStat(Pokemon attacker, Pokemon defender, Battle battle) {
         //TODO: When it says it ignores changes to defense, does that just mean stage?
         return defender.calculateDefenseIgnoreStage();
     }
